refactor(calendar): update to new conventions
