+++ packages/components/src/Calendar/CalendarClassNames.ts
@@ -1,8 +1,8 @@
 /* eslint-disable camelcase */
 import { ClassNames } from "react-day-picker"
+import calendarRangeStyles from "./CalendarRange/CalendarRange.module.scss"
+import calendarSingleStyles from "./CalendarSingle/CalendarSingle.module.scss"
 import calendarStyles from "./Calendar.module.scss"
-import calendarRangeStyles from "./CalendarRange.module.scss"
-import calendarSingleStyles from "./CalendarSingle.module.scss"
 
 const baseCalendarClasses: ClassNames = {
   vhidden: calendarStyles.vHidden,
