+++ packages/components/src/Calendar/CalendarSingle/CalendarSingle.tsx
@@ -3,10 +3,10 @@ import { enAU } from "date-fns/locale"
 import { DayPicker, DayPickerSingleProps } from "react-day-picker"
 import { ArrowBackwardIcon, ArrowForwardIcon } from "~components/Icon"
 import { OverrideClassName } from "~types/OverrideClassName"
-import { calendarSingleClasses } from "./CalendarClassNames"
-import { DayOfWeek } from "./enums"
-import { isInvalidDate } from "./utils/isInvalidDate"
-import { isValidWeekStartsOn } from "./utils/isValidWeekStartsOn"
+import { calendarSingleClasses } from "../CalendarClassNames"
+import { DayOfWeek } from "../enums"
+import { isInvalidDate } from "../utils/isInvalidDate"
+import { isValidWeekStartsOn } from "../utils/isValidWeekStartsOn"
 
 export type CalendarSingleElement = HTMLDivElement
