+++ packages/components/src/Calendar/index.ts
@@ -1,3 +1,3 @@
 export * from "./CalendarSingle"
 export * from "./CalendarRange"
-export * from "./LegacyCalendarRange"
+export * from "./LegacyCalendarRange/LegacyCalendarRange"
